Fixed networking and provisioning issues.
Network bridging was causing initial setup to stall. By switching to a
host-only network, I was able to fix this. There is no need I can see
for the VM to be visible to other machines on the network.

Regarding provisioning issues, some unnecessary networking changes were
causing errors.  I just cut them out and it seems to work fine.